Update README.md
Where did **TeamBuilder** come from? 

My open source hobby project! TeamBuilder is the name of one of the greatest world wide issues that I am developing a solution for, everyone is a part of the Team if they want to at no cost to you! It's also how I practiced Web API and the latest REST technologies during a lengthy interview process, what more could I ask for? This was the perfect study leading into the future challenges of building a Team for anything and everything possible! 

Idea for for the world's first pick up group to beat a real Professional Team (as the fun proof) of Team Builders ultimate success! Imagine being actually prepared to race for world first in new game releases? What else can you imagine is directly proportional to Team Builders progress! It's never work, it's your inspiration is the proven real contribution. You can own and inspire at the same time, the success will be how inspired each and every person (including the need for a Team Builder app/tool) we are going into a worldly challenge therefore worldly inspiration is directly proportional to the exact success/solution hence TeamBuilder's global purpose to work with ANY other TeamBuilder or roster API feed, using simply a common roster language! This proves that the more Team Builders the better if we're all on this Team together. This fun proof puts the literal capital T in Team! How cool was this to discover as a hobby and inventor mindset engineer myself AKA, and to tackle the most frustrating issue that anyone that takes any event seriously in terms of reliable fairness and competitiveness, you can take the week off and Team Builder should plug and play ANY roster for anything ever that can require a 1 or more person Team. The solution is putting the literal Capital T in Team together as this global Team we are! See the awesomeness? Here comes the inspiration! Let's go Team! Our Heroes like Richard Simmons, Billy Blanks, and the rest of our heart healthy leaders had it right the whole time! "Let's go people work it yeah!" ~Simmons fired up! You get the idea! Now just don't forget what I said generates inspiration because all we need is that capital T in Team! Yeah! see the Github for more info! We put the Capital T in Team! 
"Need a Team? The idea is to build a Team for anything and everything! Video games, sports, etc. Quick and simple. Also designed in a way where all other Team Builder apps can all work together using a common roster language!".

Github:
https://github.com/TeamBuilderApp/TeamBuilder